Move to tsdx and jest
