+++ .prettierrc
@@ -1,6 +1,5 @@
 {
-    "tabWidth": 4,
+    "tabWidth": 2,
     "semi": false,
     "singleQuote": true
-  }
-  
+}
