+++ package.json
@@ -2,29 +2,21 @@
   "name": "route-node",
   "version": "3.4.2",
   "description": "A package to create a tree of named routes",
-  "main": "dist/cjs/route-node.js",
-  "jsnext:main": "dist/es/route-node.js",
-  "module": "dist/es/route-node.js",
+  "main": "dist/index.js",
+  "module": "dist/route-node.esm.js",
+  "typings": "dist/index.d.ts",
   "sideEffects": false,
-  "typings": "./typings/RouteNode.d.ts",
+  "files": [
+    "dist",
+    "LICENSE"
+  ],
   "scripts": {
-    "test": "mocha -r ts-node/register 'test/main.js'",
-    "lint": "tslint modules/*.ts",
-    "clean": "rimraf dist",
-    "build": "npm run clean && rollup -c rollup.config.js",
-    "clog": "conventional-changelog -p angular -i CHANGELOG.md -s",
-    "precommit": "lint-staged"
-  },
-  "lint-staged": {
-    "modules/*.ts": [
-      "prettier --write",
-      "tslint",
-      "git add"
-    ],
-    "test/*.js": [
-      "prettier --write",
-      "git add"
-    ]
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test",
+    "lint": "tsdx lint src test",
+    "prepare": "tsdx build",
+    "clog": "conventional-changelog -p angular -i CHANGELOG.md -s"
   },
   "repository": {
     "type": "git",
@@ -36,30 +28,26 @@
     "route matching"
   ],
   "author": "Thomas Roch",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/troch/route-node/issues"
   },
   "homepage": "https://github.com/troch/route-node",
   "devDependencies": {
-    "conventional-changelog-cli": "~2.0.1",
-    "husky": "~0.14.3",
-    "lint-staged": "~7.0.0",
-    "lodash.omit": "~4.5.0",
-    "mocha": "~5.0.5",
-    "prettier": "~1.11.1",
-    "rimraf": "~2.6.2",
-    "rollup": "~0.57.1",
-    "rollup-plugin-node-resolve": "~3.3.0",
-    "rollup-plugin-typescript2": "~0.12.0",
-    "should": "~13.2.1",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "tslint-config-prettier": "~1.10.0",
-    "typescript": "~2.7.2"
+    "conventional-changelog-cli": "2.0.31",
+    "@types/jest": "^24.0.25",
+    "husky": "^3.1.0",
+    "tsdx": "^0.12.1",
+    "tslib": "^1.10.0",
+    "typescript": "^3.7.4"
   },
   "dependencies": {
-    "path-parser": "4.2.0",
+    "path-parser": "5.0.1",
     "search-params": "2.1.3"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint src test"
+    }
   }
 }
