--- tslint.json
@@ -1,11 +1,0 @@
-{
-    "defaultSeverity": "error",
-    "extends": ["tslint:recommended", "tslint-config-prettier"],
-    "jsRules": {},
-    "rules": {
-        "object-literal-sort-keys": false,
-        "interface-name": false,
-        "no-shadowed-variable": false
-    },
-    "rulesDirectory": []
-}
