--- typings/matchChildren.d.ts
@@ -1,3 +1,0 @@
-import RouteNode, { MatchOptions, MatchResponse } from './RouteNode';
-declare const matchChildren: (nodes: RouteNode[], pathSegment: string, currentMatch: MatchResponse, options?: MatchOptions, consumedBefore?: string) => any;
-export default matchChildren;
