--- typings/sortChildren.d.ts
@@ -1,2 +1,0 @@
-import RouteNode from './RouteNode';
-export default function sortChildren(children: RouteNode[]): RouteNode[];
